[Quartermaster] refactor: Split FeatsPanel.axaml.cs into partial classes (#1124)
Split 1,195-line file into focused partial classes:
- FeatsPanel.axaml.cs (435 LOC) - Core initialization, loading
- FeatsPanel.Search.cs (80 LOC) - Search/filter
- FeatsPanel.Display.cs (320 LOC) - Summary, assigned list, theme
- FeatsPanel.Selection.cs (99 LOC) - Add/remove feats
- FeatsPanel.SpecialAbilities.cs (172 LOC) - Special abilities

Moved ViewModels to dedicated files:
- FeatListViewModel.cs (177 LOC)
- SpecialAbilityViewModel.cs (56 LOC)

All files under 500 LOC. Tests pass ✅
